pyproject.toml
[project]
name = "aleksandrova-dacha"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.8.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-jwt>=1.11.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "notebook>=7.4.4",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "django-debug-toolbar>=6.0.0",
    "django-stubs[compatible-mypy]>=5.2.5",
    "djangorestframework-stubs>=3.16.4",
    "flake8>=7.3.0",
    "flake8-import-order>=0.19.2",
    "isort>=6.1.0",
    "mypy>=1.18.2",
]
prod = [
    "gunicorn>=23.0.0",
]
